finalização do CRUD de ClienteENtity e Create, Read e Delete de EventoEntity

+++ Buffet/Controllers/ClienteController.cs
@@ -19,19 +19,27 @@ namespace Buffet.Controllers
         private readonly ClienteService _clienteService;
         private readonly TipoClienteService _tipoClienteService;
         private readonly EnderecoService _enderecoService;
+        private readonly EventoService _eventoService;
         private readonly ILogger<ClienteController> _logger;
 
-        public ClienteController(ClienteService clienteService, ILogger<ClienteController> logger)
+        public ClienteController(ClienteService clienteService, TipoClienteService tipoClienteService, EnderecoService enderecoService, 
+            EventoService eventoService, ILogger<ClienteController> logger)
         {
             _clienteService = clienteService;
+            _tipoClienteService = tipoClienteService;
+            _enderecoService = enderecoService;
+            _eventoService = eventoService;
             _logger = logger;
-            
         }
+
+
+
 
         // GET
         public async Task<IActionResult> Index()
         {
             var clientes = await _clienteService.getAll();
+            
             var clienteViewModel = new ClienteViewModel();
             clienteViewModel.Clientes = clientes;
             return View("~/Views/Private/ListaCliente.cshtml", clienteViewModel);
@@ -42,21 +50,32 @@ namespace Buffet.Controllers
             var cvm = new ClienteViewModel();
             return View("~/Views/Private/CadastroCliente.cshtml", cvm);
         }
+        
+        public async Task<IActionResult> ClientesFiltroAsync(string buscaNome, string buscaEmail)
+        {
+
+            var cvm = new ClienteViewModel();
+            cvm.Clientes = await _clienteService.buscaClientes(buscaNome, buscaEmail);
+            return View("~/Views/Private/ListaCliente.cshtml", cvm);
+        }
 
 
         public async Task<IActionResult> Store(int id, string nomeCliente, string tipoCliente, 
                                                 string emailCliente, string enderecoRuaCliente, string enderecoBairroCliente, string enderecoEstadoCliente,
-                                                string enderecoCidadeCliente, int enderecoNumCliente, string obsCliente, DateTime criadoEm)
+                                                string enderecoCidadeCliente, int enderecoNumCliente, string obsCliente)
         {
             if (id == 0)
             {
-                criadoEm = DateTime.Today;
+                var criadoEm = DateTime.Today;
                 await _clienteService.store(nomeCliente, tipoCliente, emailCliente, enderecoRuaCliente, enderecoBairroCliente, enderecoEstadoCliente, 
                                             enderecoCidadeCliente, enderecoNumCliente, obsCliente, criadoEm);
             }
             else
             {
-                //await _clienteService.update(id);
+                var editadoEm = DateTime.Today;
+                var eventos = _eventoService.getEventosByCliente(id);
+                await _clienteService.update(id, nomeCliente, tipoCliente, emailCliente, enderecoRuaCliente, enderecoBairroCliente, enderecoEstadoCliente, 
+                    enderecoCidadeCliente, enderecoNumCliente, obsCliente, editadoEm, eventos);
             }
             return RedirectToAction("Index");
         }
@@ -64,19 +83,37 @@ namespace Buffet.Controllers
         public async Task<IActionResult> Edit(int id)
         {
             var cliente = await _clienteService.getById(id);
-            EnderecoEntity endereco = await _enderecoService.getById(cliente.Endereco.Id);
-            TipoClienteEntity tipo = await _tipoClienteService.getById(cliente.TipoCliente.Id);
-            var clienteViewModel = new ClienteViewModel(id, cliente.Nome,
-                tipo.Descricao, cliente.Email, endereco.Estado, endereco.Cidade, endereco.Bairro, endereco.Rua, endereco.Numero, cliente.Observacoes);
-            return View("~/Views/Private/CadastroCliente.cshtml", clienteViewModel);
+            var enderecoCliente = await _enderecoService.getById(cliente.Endereco.Id);
+            var lvm = new ClienteViewModel()
+            {
+                id = cliente.Id,
+                nome = cliente.Nome,
+                email = cliente.Email,
+                Clientes = await _clienteService.getAll(),
+                Bairro = enderecoCliente.Bairro,
+                Cidade = enderecoCliente.Cidade,
+                obs = cliente.Observacoes,
+                Estado = enderecoCliente.Estado,
+                Numero = enderecoCliente.Numero,
+                Rua = enderecoCliente.Rua
+            };
+            return View("~/Views/Private/CadastroCliente.cshtml", lvm);
         }
 
         public async Task<IActionResult> Destroy(int id)
         {
-            await _clienteService.destroy(id);
             var clientes = await _clienteService.getAll();
+            var cliente = await _clienteService.getById(id);
             var clienteViewModel = new ClienteViewModel();
-            clienteViewModel.mensagem = "Deletado com sucesso!";
+            if(cliente.Eventos.Capacity > 0)
+            {
+                clienteViewModel.mensagem = "Cliente com evento agendado não pode ser excluído!";
+                clienteViewModel.Clientes = clientes;
+            } else
+            {
+                await _clienteService.destroy(id);
+                clienteViewModel.mensagem = "Deletado com sucesso!";
+            }
             clienteViewModel.Clientes = clientes;
             return View("~/Views/Private/ListaCliente.cshtml", clienteViewModel);
         }
